WWH/ATS Resume Checker for enhancing job opportunities for women/requirements.txt
# streamlit
# google-generativeai
# python-dotenv
# pdf2image
streamlit
PyPDF2
google.generativeai 
python-dotenv